+++ crates/fuel-core/src/schema/coins.rs
@@ -4,8 +4,8 @@ use crate::{
         SpendQuery,
     },
     fuel_core_graphql_api::{
+        query_costs,
         IntoApiResult,
-        QUERY_COSTS,
     },
     graphql_api::api_service::ConsensusProvider,
     query::asset_query::AssetSpendTarget,
@@ -93,7 +93,7 @@ impl MessageCoin {
         self.0.amount.into()
     }
 
-    #[graphql(complexity = "QUERY_COSTS.storage_read")]
+    #[graphql(complexity = "query_costs().storage_read")]
     async fn asset_id(&self, ctx: &Context<'_>) -> AssetId {
         let params = ctx
             .data_unchecked::<ConsensusProvider>()
@@ -149,7 +149,7 @@ pub struct CoinQuery;
 #[async_graphql::Object]
 impl CoinQuery {
     /// Gets the coin by `utxo_id`.
-    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    #[graphql(complexity = "query_costs().storage_read + child_complexity")]
     async fn coin(
         &self,
         ctx: &Context<'_>,
@@ -161,9 +161,9 @@ impl CoinQuery {
 
     /// Gets all unspent coins of some `owner` maybe filtered with by `asset_id` per page.
     #[graphql(complexity = "{\
-        QUERY_COSTS.storage_iterator\
-        + (QUERY_COSTS.storage_read + first.unwrap_or_default() as usize) * child_complexity \
-        + (QUERY_COSTS.storage_read + last.unwrap_or_default() as usize) * child_complexity\
+        query_costs().storage_iterator\
+        + (query_costs().storage_read + first.unwrap_or_default() as usize) * child_complexity \
+        + (query_costs().storage_read + last.unwrap_or_default() as usize) * child_complexity\
     }")]
     async fn coins(
         &self,
@@ -206,7 +206,7 @@ impl CoinQuery {
     ///     The list of spendable coins per asset from the query. The length of the result is
     ///     the same as the length of `query_per_asset`. The ordering of assets and `query_per_asset`
     ///     is the same.
-    #[graphql(complexity = "QUERY_COSTS.coins_to_spend")]
+    #[graphql(complexity = "query_costs().coins_to_spend")]
     async fn coins_to_spend(
         &self,
         ctx: &Context<'_>,