+++ crates/fuel-core/src/schema/message.rs
@@ -11,7 +11,7 @@ use super::{
     ReadViewProvider,
 };
 use crate::{
-    fuel_core_graphql_api::QUERY_COSTS,
+    fuel_core_graphql_api::query_costs,
     graphql_api::IntoApiResult,
     schema::scalars::{
         BlockId,
@@ -66,7 +66,7 @@ pub struct MessageQuery {}
 
 #[Object]
 impl MessageQuery {
-    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    #[graphql(complexity = "query_costs().storage_read + child_complexity")]
     async fn message(
         &self,
         ctx: &Context<'_>,
@@ -78,9 +78,9 @@ impl MessageQuery {
     }
 
     #[graphql(complexity = "{\
-        QUERY_COSTS.storage_iterator\
-        + (QUERY_COSTS.storage_read + first.unwrap_or_default() as usize) * child_complexity \
-        + (QUERY_COSTS.storage_read + last.unwrap_or_default() as usize) * child_complexity\
+        query_costs().storage_iterator\
+        + (query_costs().storage_read + first.unwrap_or_default() as usize) * child_complexity \
+        + (query_costs().storage_read + last.unwrap_or_default() as usize) * child_complexity\
     }")]
     async fn messages(
         &self,
@@ -128,7 +128,7 @@ impl MessageQuery {
     }
 
     // 256 * QUERY_COSTS.storage_read because the depth of the Merkle tree in the worst case is 256
-    #[graphql(complexity = "256 * QUERY_COSTS.storage_read + child_complexity")]
+    #[graphql(complexity = "256 * query_costs().storage_read + child_complexity")]
     async fn message_proof(
         &self,
         ctx: &Context<'_>,
@@ -159,7 +159,7 @@ impl MessageQuery {
         .map(MessageProof))
     }
 
-    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    #[graphql(complexity = "query_costs().storage_read + child_complexity")]
     async fn message_status(
         &self,
         ctx: &Context<'_>,