+++ crates/fuel-core/src/schema/tx.rs
@@ -6,8 +6,8 @@ use crate::{
             ConsensusProvider,
             TxPool,
         },
+        query_costs,
         IntoApiResult,
-        QUERY_COSTS,
     },
     graphql_api::{
         database::ReadView,
@@ -87,7 +87,7 @@ pub struct TxQuery;
 
 #[Object]
 impl TxQuery {
-    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    #[graphql(complexity = "query_costs().storage_read + child_complexity")]
     async fn transaction(
         &self,
         ctx: &Context<'_>,
@@ -109,7 +109,7 @@ impl TxQuery {
 
     // We assume that each block has 100 transactions.
     #[graphql(complexity = "{\
-        (QUERY_COSTS.tx_get + child_complexity) \
+        (query_costs().tx_get + child_complexity) \
         * (first.unwrap_or_default() as usize + last.unwrap_or_default() as usize)
     }")]
     async fn transactions(
@@ -193,9 +193,9 @@ impl TxQuery {
     }
 
     #[graphql(complexity = "{\
-        QUERY_COSTS.storage_iterator\
-        + (QUERY_COSTS.storage_read + first.unwrap_or_default() as usize) * child_complexity \
-        + (QUERY_COSTS.storage_read + last.unwrap_or_default() as usize) * child_complexity\
+        query_costs().storage_iterator\
+        + (query_costs().storage_read + first.unwrap_or_default() as usize) * child_complexity \
+        + (query_costs().storage_read + last.unwrap_or_default() as usize) * child_complexity\
     }")]
     async fn transactions_by_owner(
         &self,
@@ -237,7 +237,7 @@ impl TxQuery {
     }
 
     /// Estimate the predicate gas for the provided transaction
-    #[graphql(complexity = "QUERY_COSTS.estimate_predicates + child_complexity")]
+    #[graphql(complexity = "query_costs().estimate_predicates + child_complexity")]
     async fn estimate_predicates(
         &self,
         ctx: &Context<'_>,
@@ -282,7 +282,7 @@ pub struct TxMutation;
 impl TxMutation {
     /// Execute a dry-run of multiple transactions using a fork of current state, no changes are committed.
     #[graphql(
-        complexity = "QUERY_COSTS.dry_run * txs.len() + child_complexity * txs.len()"
+        complexity = "query_costs().dry_run * txs.len() + child_complexity * txs.len()"
     )]
     async fn dry_run(
         &self,
@@ -334,7 +334,7 @@ impl TxMutation {
     /// Submits transaction to the `TxPool`.
     ///
     /// Returns submitted transaction if the transaction is included in the `TxPool` without problems.
-    #[graphql(complexity = "QUERY_COSTS.submit + child_complexity")]
+    #[graphql(complexity = "query_costs().submit + child_complexity")]
     async fn submit(
         &self,
         ctx: &Context<'_>,
@@ -374,7 +374,7 @@ impl TxStatusSubscription {
     /// then the updates arrive. In such a case the stream will close without
     /// a status. If this occurs the stream can simply be restarted to return
     /// the latest status.
-    #[graphql(complexity = "QUERY_COSTS.status_change + child_complexity")]
+    #[graphql(complexity = "query_costs().status_change + child_complexity")]
     async fn status_change<'a>(
         &self,
         ctx: &'a Context<'a>,
@@ -394,7 +394,7 @@ impl TxStatusSubscription {
     }
 
     /// Submits transaction to the `TxPool` and await either confirmation or failure.
-    #[graphql(complexity = "QUERY_COSTS.submit_and_await + child_complexity")]
+    #[graphql(complexity = "query_costs().submit_and_await + child_complexity")]
     async fn submit_and_await<'a>(
         &self,
         ctx: &'a Context<'a>,
@@ -413,7 +413,7 @@ impl TxStatusSubscription {
     /// Submits the transaction to the `TxPool` and returns a stream of events.
     /// Compared to the `submitAndAwait`, the stream also contains `
     /// SubmittedStatus` as an intermediate state.
-    #[graphql(complexity = "QUERY_COSTS.submit_and_await + child_complexity")]
+    #[graphql(complexity = "query_costs().submit_and_await + child_complexity")]
     async fn submit_and_await_status<'a>(
         &self,
         ctx: &'a Context<'a>,