+++ tests/tests/dos.rs
@@ -237,7 +237,7 @@ async fn complex_queries__40_full_blocks__works() {
 }
 
 #[tokio::test]
-async fn complex_queries__41_full_block__query_to_complex() {
+async fn complex_queries__41_full_block__query_too_complex() {
     let query = FULL_BLOCK_QUERY.to_string();
     let query = query.replace("$NUMBER_OF_BLOCKS", "41");
 
@@ -246,6 +246,21 @@ async fn complex_queries__41_full_block__query_to_complex() {
 
     let result = send_graph_ql_query(&url, query.as_str()).await;
     assert!(result.contains("Query is too complex."));
+}
+
+#[tokio::test]
+async fn complex_queries__increased_block_header_cost__failed_to_initialize_service() {
+    let mut config = Config::local_node();
+    config.graphql_config.costs.block_header =
+        config.graphql_config.max_queries_complexity;
+
+    let Err(error) = FuelService::new_node(config).await else {
+        panic!("expected error");
+    };
+
+    assert!(error
+        .to_string()
+        .contains("cannot initialize queries with non-default costs in tests"))
 }
 
 #[tokio::test]